+++ lighthouse-cli/test/cli/printer-test.js
@@ -24,8 +24,8 @@ describe('Printer', () => {
 
   it('writes file for results', () => {
     const path = './.test-file.json';
-    const formattedReport = JSON.stringify(sampleResults);
-    return Printer.write({formattedReport}, 'json', path).then(_ => {
+    const report = JSON.stringify(sampleResults);
+    return Printer.write({report}, 'json', path).then(_ => {
       const fileContents = fs.readFileSync(path, 'utf8');
       assert.ok(/lighthouseVersion/gim.test(fileContents));
       fs.unlinkSync(path);
@@ -34,8 +34,8 @@ describe('Printer', () => {
 
   it('throws for invalid paths', () => {
     const path = '!/#@.json';
-    const formattedReport = JSON.stringify(sampleResults);
-    return Printer.write({formattedReport}, 'html', path).catch(err => {
+    const report = JSON.stringify(sampleResults);
+    return Printer.write({report}, 'html', path).catch(err => {
       assert.ok(err.code === 'ENOENT');
     });
   });
