--- lighthouse-core/report/v2/report-formatter.js
@@ -1,33 +1,0 @@
-/**
- * @license Copyright 2018 Google Inc. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
- */
-'use strict';
-
-const ReportGenerator = require('./report-generator');
-
-module.exports = {
-  /**
-   * Creates the results output in a format based on the `mode`.
-   * @param {LH.Result} lhr
-   * @param {string} outputMode
-   * @return {string}
-   */
-  formatReport(lhr, outputMode) {
-    // HTML report.
-    if (outputMode === 'html') {
-      return ReportGenerator.generateReportHtml(lhr);
-    }
-    // CSV report.
-    if (outputMode === 'csv') {
-      return ReportGenerator.generateReportCSV(lhr);
-    }
-    // JSON report.
-    if (outputMode === 'json') {
-      return JSON.stringify(lhr, null, 2);
-    }
-
-    throw new Error('Invalid output mode: ' + outputMode);
-  },
-};
