+++ lighthouse-core/test/report/v2/report-generator-test.js
@@ -10,6 +10,8 @@ const fs = require('fs');
 const jsdom = require('jsdom');
 const ReportGeneratorV2 = require('../../../report/v2/report-generator.js');
 const TEMPLATES_FILE = fs.readFileSync(__dirname + '/../../../report/v2/templates.html', 'utf8');
+const sampleResults = require('../../../../lighthouse-core/test/results/sample_v2.json');
+const csvValidator = require('csv-validator');
 
 /* eslint-env mocha */
 
@@ -69,6 +71,47 @@ describe('ReportGeneratorV2', () => {
       assert.ok(result.includes('robustness: \\u003c/script'), 'escapes HTML tags in javascript');
       assert.ok(result.includes('pre$`post'), 'does not break from String.replace');
       assert.ok(result.includes('LIGHTHOUSE_JSON'), 'cannot be tricked');
+    });
+  });
+
+  describe('#generateReport', () => {
+    it('creates JSON for results', () => {
+      const jsonOutput = ReportGeneratorV2.generateReport(sampleResults, 'json');
+      assert.doesNotThrow(_ => JSON.parse(jsonOutput));
+    });
+
+    it('creates HTML for results', () => {
+      const htmlOutput = ReportGeneratorV2.generateReport(sampleResults, 'html');
+      assert.ok(/<!doctype/gim.test(htmlOutput));
+      assert.ok(/<html lang="en"/gim.test(htmlOutput));
+    });
+
+    it('creates CSV for results', async () => {
+      const path = './.results-as-csv.csv';
+      const headers = {
+        category: '',
+        name: '',
+        title: '',
+        type: '',
+        score: 42,
+      };
+
+      const csvOutput = ReportGeneratorV2.generateReport(sampleResults, 'csv');
+      fs.writeFileSync(path, csvOutput);
+
+      try {
+        await csvValidator(path, headers);
+      } catch (err) {
+        assert.fail('CSV parser error:\n' + err.join('\n'));
+      } finally {
+        fs.unlinkSync(path);
+      }
+    });
+
+    it('writes extended info', () => {
+      const htmlOutput = ReportGeneratorV2.generateReport(sampleResults, 'html');
+      const outputCheck = new RegExp('dobetterweb/dbw_tester.css', 'i');
+      assert.ok(outputCheck.test(htmlOutput));
     });
   });
 });
